Re-add Generated_Source directory and remove from gitignore
The name "Generated_Source" is misleading - it is not 100% autogenerated
code, but contains code that is added by designers and by the user.
This is critical code to be version controlled unlike normal generated
source. This commit adds back in the code removed when the rest of the
generated code was removed from the repo